% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redact_all.R
\name{default_redact_functions}
\alias{default_redact_functions}
\title{Get a list of all redact functions in the package}
\usage{
default_redact_functions()
}
\value{
A character vector with the names of all redact functions.
}
\description{
This function returns a character vector with the names of all redact functions
defined in the redacted package.
}
\examples{
default_redact_functions()

}
\keyword{functions}
\keyword{redact}
