% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redact_credit_card_numbers.R
\name{redact_credit_card_numbers}
\alias{redact_credit_card_numbers}
\title{Redact credit card numbers in text}
\usage{
redact_credit_card_numbers(text, replace_with = NULL)
}
\arguments{
\item{text}{The text in which credit card numbers are to be redacted.}

\item{replace_with}{A string to replace the credit card numbers with.
It can either be a whole word like "[REDACTED]", or a single character like "*"
that gets repeated based on how long the credit card number detected is.
Default is NULL, which uses the default value specified in the
`replace_pattern()` function.}
}
\value{
A character vector with credit card numbers redacted.
}
\description{
This function redacts credit card numbers in a given text by replacing them
  with a specified string.
}
\examples{
text <- "My credit card number is 1234 5678 9012 3456. Please don't share it with anyone."
redacted_text <- redact_credit_card_numbers(text)
cat(redacted_text)

}
