% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redact_email_addresses.R
\name{redact_email_addresses}
\alias{redact_email_addresses}
\title{Redact email addresses in text}
\usage{
redact_email_addresses(text, replace_with = NULL)
}
\arguments{
\item{text}{The text in which email addresses are to be redacted.}

\item{replace_with}{A string to replace email addresses with.
It can be either a whole word pattern or a single character.
Default is NULL, which uses the default value specified in the
`replace_pattern()` function.}
}
\value{
A character vector with redacted email addresses.
}
\description{
This function redacts email addresses in a given text by replacing them with a specified string.
}
\examples{
text <- "Hello, my email address is john.doe@gmail.com. Please don't share it with anyone."
redacted_text <- redact_email_addresses(text)
cat(redacted_text)

}
