% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redact_ip_addresses.R
\name{redact_ip_addresses}
\alias{redact_ip_addresses}
\title{Redact IP addresses in text}
\usage{
redact_ip_addresses(text, replace_with = NULL)
}
\arguments{
\item{text}{The text in which IP addresses are to be redacted.}

\item{replace_with}{A string to replace IP addresses with.
It can be either a whole word pattern or a single character.
Default is NULL, which uses the default value specified in the
`replace_pattern()` function.}
}
\value{
A character vector with redacted IP addresses.
}
\description{
This function redacts IP addresses in a given text by replacing them with a specified string.
}
\examples{
text <- "My IP address is 192.168.0.1. Please don't share it with anyone."
redacted_text <- redact_ip_addresses(text)
cat(redacted_text)

}
