% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redact_phone_numbers.R
\name{redact_phone_numbers}
\alias{redact_phone_numbers}
\title{Redact phone numbers in text}
\usage{
redact_phone_numbers(text, replace_with = NULL)
}
\arguments{
\item{text}{The text in which phone numbers are to be redacted.}

\item{replace_with}{A string to replace the phone numbers with.
It can either be a whole word like "[REDACTED]", or a single character like "*"
that gets repeated based on how long the phone number detected is.
Default is NULL, which uses the default value specified in the
`replace_pattern()` function.}
}
\value{
A character vector with phone numbers redacted.
}
\description{
This function redacts phone numbers in a given text by replacing them with a specified string.
}
\examples{
text <- "My phone number is (123) 456-7890. Please don't share it with anyone."
redacted_text <- redact_phone_numbers(text)
cat(redacted_text)

}
