% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redact_pins.R
\name{redact_pins}
\alias{redact_pins}
\title{Redact 4 digit pins from a string}
\usage{
redact_pins(text, replace_with = NULL)
}
\arguments{
\item{text}{A character string.}

\item{replace_with}{A character string to replace the matched pattern with.
If NULL, defaults to "[REDACTED]".}
}
\value{
The input string with any 4 digit pins replaced with [REDACTED].
}
\description{
This function takes a string and replaces any occurrences of a 4 digit pin
}
\examples{
redact_pins("My pin is 1234")
redact_pins("Her PIN is 9876")

}
