% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redact_social_security_numbers.R
\name{redact_social_security_numbers}
\alias{redact_social_security_numbers}
\title{Redact social security numbers in text}
\usage{
redact_social_security_numbers(text, replace_with = NULL)
}
\arguments{
\item{text}{The text in which social security numbers are to be redacted.}

\item{replace_with}{A string to replace the social security numbers with.
It can either be a whole word like "[REDACTED]", or a single character like "*"
that gets repeated based on how long the social security number detected is.
Default is NULL, which uses the default value specified in the
`replace_pattern()` function.}
}
\value{
A character vector with social security numbers redacted.
}
\description{
This function redacts social security numbers in a given text by replacing them with a specified string.
}
\examples{
text <- "My social security number is 123-45-6789. Please don't share it with anyone."
redacted_text <- redact_social_security_numbers(text)
cat(redacted_text)

}
