% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redact_website_urls.R
\name{redact_website_urls}
\alias{redact_website_urls}
\title{Redact website URLs in a text string}
\usage{
redact_website_urls(text, replace_with = NULL)
}
\arguments{
\item{text}{A character string containing the text to be searched and redacted}

\item{replace_with}{A character string that specifies the value to replace the
matched domains with. If this argument is not specified, the default value of
"[REDACTED]" will be used.}
}
\value{
A character string with matched domains replaced with the specified or default value
}
\description{
This function searches a text string for website URLs and replaces them with
a specified value or the default value of "[REDACTED]".
}
\examples{
text <- "Visit our website at www.example.com for more information."
redact_website_urls(text)

}
