% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redact_pattern.R
\name{replace_pattern}
\alias{replace_pattern}
\title{Replace pattern in text}
\usage{
replace_pattern(text, pattern, replace_with = "[REDACTED]")
}
\arguments{
\item{text}{The text in which the pattern is to be replaced.}

\item{pattern}{The pattern to replace in the text.}

\item{replace_with}{A string to replace the pattern with.}
}
\value{
A character vector with the pattern replaced.
}
\description{
This function replaces a specified pattern in a given text with a specified string.
}
